requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# Database
supabase==2.3.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Machine Learning
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.25.2
tensorflow==2.15.0
joblib==1.3.2

# RAG and AI
openai==1.3.7
langchain==0.0.350
langchain-community==0.0.10
chromadb==0.4.18
sentence-transformers==2.2.2

# Notifications
requests==2.31.0
smtplib-ssl==1.0.0
slack-sdk==3.26.1

# Monitoring and Logging
prometheus-client==0.19.0
structlog==23.2.0

# Security
cryptography==41.0.8
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
